+++ code/inetfile/cftp.cpp
@@ -143,7 +143,7 @@ CFtpGet::CFtpGet(char *URL, char *localfile, char *Username, char *Password)
 	//when you found it, you have the host and dir
 	char *filestart = NULL;
 	char *dirstart = NULL;
-	for(int i = strlen(pURL);i>=0;i--)
+	for(size_t i = strlen(pURL);i>=0;i--)
 	{
 		if(pURL[i]== '/')
 		{
